pyproject.toml
[project]
name = "image-to-ascii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pillow>=11.3.0",
]
